% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_house_mortgage.R
\name{calc_house_mortgage}
\alias{calc_house_mortgage}
\title{Calculate house mortgage}
\usage{
calc_house_mortgage(
  principal,
  interest_rate,
  period_year = 10,
  initial_cost,
  start_date = "2023-04-25",
  payment_method = "even principal",
  verbose = FALSE,
  plotit = TRUE
)
}
\arguments{
\item{principal}{Numeric value of principal}

\item{interest_rate}{Numeric value of annual interest rate}

\item{period_year}{Numeric value of period of mortgage}

\item{initial_cost}{Numeric value of initial cost}

\item{start_date}{Character value of starting date}

\item{payment_method}{Character value (Default: "even principal") indicating a method of payment: either "even principal" or "even total"}

\item{verbose}{Logical value (Default: FALSE) indicating whether you want to see verbose}

\item{plotit}{Logical value (Default: TRUE) indicating whether you want to plot the ggplot}
}
\value{
Data frame and ggplot2 object
}
\description{
Calculate house mortgage
}
