% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stockJ_data.R
\name{get_stockJ_data}
\alias{get_stockJ_data}
\title{Get stock price history in Japan based on Yahoo! Finance webpage}
\usage{
get_stockJ_data(code)
}
\arguments{
\item{code}{Numeric value representing the stock ID number that you want to search}
}
\value{
Dataframe
}
\description{
This function scrapes the stock price information from the Yahoo! finance webpage (https://finance.yahoo.co.jp).
}
