% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_US_FRM_rate.R
\name{get_US_FRM_rate}
\alias{get_US_FRM_rate}
\title{Get U.S. mortgage rate information from Freddie Mac website}
\usage{
get_US_FRM_rate(plotit = TRUE, st_date = NULL, ed_date = NULL)
}
\arguments{
\item{plotit}{Logical value (Default: TRUE) indicating whether or not you want to plot the rate time-series.}

\item{st_date}{Character value (Default: NULL) indicating a starting date}

\item{ed_date}{Character value (Default: NULL) indicating a ending date}
}
\value{
Data frame (and ggplot object)
}
\description{
The mortgage rate data is taken from Federal Home Loan Mortgage Corporation (FHLMC), commonly known as Freddie Mac.
}
