% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_saison_csv.R
\name{read_saison_csv}
\alias{read_saison_csv}
\title{Load SAISON credit card history data from csv}
\usage{
read_saison_csv()
}
\value{
Data frame
}
\description{
Load SAISON credit card history data from csv
}
