% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prime_rate.R
\name{get_prime_rate}
\alias{get_prime_rate}
\title{Get prime rate information from Bank of Japan website (https://www.boj.or.jp/statistics/dl/loan/prime/prime.htm)}
\usage{
get_prime_rate(plotit = TRUE)
}
\arguments{
\item{plotit}{Logical value (Default: TRUE) indicating whether you want to plot the result}
}
\value{
Data frame (and ggplot object)
}
\description{
Get prime rate information from Bank of Japan website (https://www.boj.or.jp/statistics/dl/loan/prime/prime.htm)
}
